==> sfi-asset-level-data-develop/phase-2-deployment/src/main/jupyter/06-s2-rgb-macro-localization-model-deployment/00-s2-rgb-aws-management.ipynb <==
# View assets on S3

import boto3
import os
import shutil

! pip install awscli

s3 = boto3.resource('s3')
bucket = s3.Bucket('sfi-shared-assets')

! aws s3 ls s3://sfi-shared-assets/

## Sentinel-2 Macro-localization Model Build Chips

! aws s3 ls s3://sfi-shared-assets/S2-RGB-macro-localization-model-deployment4/

! aws s3 ls s3://sfi-shared-assets/S2-RGB-macro-localization-model-build4/S2-RGB-model-results4/

## Download

bucket.download_file('S2-RGB-macro-localization-model-deployment4/S2-deployment-chip-scores-CHN-10km-nowater-2020-consolidated.geojson',
                     'S2-deployment-chip-scores-CHN-10km-nowater-2020-consolidated.geojson')

## Remove from S3?

! aws s3 rm s3://sfi-shared-assets/L8-TIR-macro-localization-model-build2/ --recursive
